# specs/001-rag-chatbot/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RAG Chatbot API
  description: |
    API for the book-grounded RAG chatbot with personalization and translation features.
    All chat responses are derived strictly from indexed book content.
  version: 1.0.0
  contact:
    name: RAG Chatbot Team
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourbook.com
    description: Production server
...
paths:
...
  /api/chat:
    post:
      tags: [Chat]
      summary: Send a chat message (full book mode)
      description: |
        Send a question to the chatbot. Answers are derived from indexed book content only.
        If the question cannot be answered from the book, returns "Not covered in this book".
      operationId: sendChatMessage
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chat/selected:
    post:
      tags: [Chat]
      summary: Send a chat message (selected text mode)
      description: |
        Send a question about selected text. Answers are restricted to the provided text only.
        Does not use RAG retrieval - uses only the selected_text as context.
      operationId: sendSelectedTextMessage
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectedTextRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request (e.g., missing selected text)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chat/conversations:
    get:
      tags: [Chat]
      summary: Get user's conversation history
      description: Returns list of conversations for authenticated user
      operationId: getConversations
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chat/conversations/{conversationId}:
    get:
      tags: [Chat]
      summary: Get conversation details with messages
      description: Returns a single conversation with all its messages
      operationId: getConversation
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetailResponse'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
          description: The user's question
        conversation_id:
          type: string
          format: uuid
          description: Optional conversation ID to continue existing conversation
    SelectedTextRequest:
      type: object
      required: [message, selected_text]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
          description: The user's question about the selected text
        selected_text:
          type: string
          minLength: 10
          maxLength: 10000
          description: The text selected by the user
        conversation_id:
          type: string
          format: uuid
          description: Optional conversation ID
    ChatResponse:
      type: object
      required: [message, conversation_id, sources]
      properties:
        message:
          type: string
          description: The assistant's response
        conversation_id:
          type: string
          format: uuid
          description: Conversation ID for follow-up messages
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceReference'
          description: Source references (empty for selected-text mode or "not covered")
        is_covered:
          type: boolean
          description: Whether the question was answerable from book content
    SourceReference:
      type: object
      required: [chapter, relevance]
      properties:
        chapter:
          type: string
          description: Chapter slug
        section:
          type: string
          description: Section title within chapter
        relevance:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Relevance score
    ConversationListResponse:
      type: object
      required: [conversations, total]
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/ConversationSummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    ConversationSummary:
      type: object
      required: [id, mode, created_at, message_count]
      properties:
        id:
          type: string
          format: uuid
        mode:
          type: string
          enum: [full_book, selected_text]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        message_count:
          type: integer
        preview:
          type: string
          description: First user message (truncated)
    ConversationDetailResponse:
      type: object
      required: [id, mode, messages, created_at]
      properties:
        id:
          type: string
          format: uuid
        mode:
          type: string
          enum: [full_book, selected_text]
        selected_text:
          type: string
          description: Only present if mode is selected_text
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        source_refs:
          type: array
          items:
            $ref: '#/components/schemas/SourceReference'
        created_at:
          type: string
          format: date-time

    # Translation
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - NOT_FOUND
                - RATE_LIMITED
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: string
              description: Additional error details (dev mode only)